move contextual sidebar to sessions' +page.svelte
