+++ src/routes/+page.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
+	import { page } from '$app/stores';
 	import { slide } from 'svelte/transition';
 
 	import Badge from '$lib/components/Badge.svelte';
@@ -8,22 +9,31 @@
 	import FieldModels from '$lib/components/FieldModels.svelte';
 	import Field from '$lib/components/Field.svelte';
 	import type { OllamaTagResponse } from '$lib/ollama';
-	import { settingsStore } from '$lib/store';
+	import { sessionsStore, settingsStore } from '$lib/store';
+	import { Settings } from 'lucide-svelte';
+	import { generateStorageId } from '$lib/utils';
 
 	export let ollamaURL: URL | null = null;
-	
+
 	const DETAULT_OLLAMA_SERVER = 'http://localhost:11434';
 	let serverStatus: 'connected' | 'disconnected' = 'connected';
 
 	let ollamaServer = $settingsStore?.ollamaServer || DETAULT_OLLAMA_SERVER;
 	let ollamaModel = $settingsStore?.ollamaModel || '';
 	let ollamaTagResponse: OllamaTagResponse | null = null;
+	let newSessionId: string;
 
 	$: settingsStore.set({
 		ollamaServer,
 		ollamaModels: ollamaTagResponse?.models || [],
 		ollamaModel
 	});
+
+	function setSessionId() {
+		newSessionId = generateStorageId();
+	}
+
+	onMount(setSessionId);
 
 	async function getModelsList(): Promise<void> {
 		try {
@@ -65,6 +75,64 @@
 		await getModelsList();
 	});
 </script>
+
+<aside class="flex h-screen flex-col">
+	<div class="flex gap-x-2 p-6">
+		<Button title="Settings" variant="outline" size="icon" href="/">
+			<Settings class="h-4 w-4" />
+		</Button>
+		<Button
+			data-testid="new-session"
+			class="w-full"
+			variant="outline"
+			href={`/${newSessionId}`}
+			on:click={setSessionId}
+		>
+			New session
+		</Button>
+	</div>
+
+	<Separator />
+
+	<div class="flex h-full flex-col overflow-y-auto py-3">
+		{#if $sessionsStore && $sessionsStore.length > 0}
+			<!-- Using slice() to reverse $sessionsStore without affecting the original array -->
+			{#each $sessionsStore.slice().reverse() as session (session.id)}
+				<a
+					data-testid="session-item"
+					href={`/${session.id}`}
+					class="
+						flex
+						flex-col
+						px-6
+						py-3
+						{$page.url.pathname.includes(session.id) ? 'bg-accent' : 'hover:bg-accent'}
+					"
+					aria-label={`Session ${session.id}`}
+					transition:slide
+				>
+					<p class="max-w-full truncate whitespace-nowrap text-sm font-bold text-foreground">
+						{session.messages[0].content}
+					</p>
+					<p
+						class="flex max-w-full gap-x-2 truncate whitespace-nowrap text-sm text-muted-foreground"
+					>
+						{session.model}
+					</p>
+				</a>
+			{/each}
+		{:else}
+			<p
+				transition:slide
+				class="flex h-full w-full items-center justify-center text-sm text-muted-foreground"
+			>
+				No sessions in history
+			</p>
+		{/if}
+	</div>
+</aside>
+
+<Separator orientation="vertical" />
 
 <div class="flex w-full flex-col bg-secondary">
 	<div class="justify-content-center m-auto max-w-[40ch] p-6">
